test(core): resucitar y podar integration_engine_tests
