--- resurrection_ward/integration_engine_tests.rs
@@ -1,290 +1,0 @@
-//! Integration tests for EngineOptions and crawl_site_with_options.
-//!
-//! Tests use wiremock for deterministic HTTP mocking — no network required.
-//!
-//! Run with: `cargo test --test integration_engine_tests`
-
-use webfang::application::crawler::engine::EngineOptions;
-use webfang::domain::JsStrategy;
-use webfang::{crawl_site_with_options, CrawlerConfig};
-use tempfile::TempDir;
-use url::Url;
-use wiremock::matchers::{method, path};
-use wiremock::{Mock, MockServer, ResponseTemplate};
-
-/// Helper: build a minimal CrawlerConfig pointing at the mock server.
-fn test_config(base_url: &str) -> CrawlerConfig {
-    let seed = Url::parse(&format!("{base_url}/index.html")).expect("valid mock URL");
-    CrawlerConfig::builder(seed)
-        .max_depth(0)
-        .max_pages(5)
-        .delay_ms(1)
-        .concurrency(1)
-        .timeout_secs(5)
-        .build()
-}
-
-/// Test 1: Engine with checkpoint enabled creates a checkpoint file.
-#[ignore = "resurrected: pending triage"]
-#[tokio::test]
-async fn test_engine_with_checkpoint_enabled() {
-    let server = MockServer::start().await;
-
-    Mock::given(method("GET"))
-        .and(path("/index.html"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_string("<html><body><h1>Hello</h1></body></html>"),
-        )
-        .mount(&server)
-        .await;
-
-    let tmp = TempDir::new().unwrap();
-    let checkpoint_dir = tmp.path().join("checkpoints");
-
-    let config = test_config(&server.uri());
-    let options = EngineOptions {
-        checkpoint_path: Some(checkpoint_dir.clone()),
-        session_pool_enabled: false,
-        ignore_robots: true,
-        js_strategy: JsStrategy::Static,
-        autoscale_enabled: false,
-        ..Default::default()
-    };
-
-    let result = crawl_site_with_options(config, options).await;
-    assert!(result.is_ok(), "crawl should succeed: {:?}", result.err());
-
-    let crawl_result = result.unwrap();
-    assert!(
-        crawl_result.total_pages >= 1,
-        "should crawl at least 1 page"
-    );
-
-    // Checkpoint file should exist after crawl
-    let checkpoint_file = checkpoint_dir.join("crawl_checkpoint.json");
-    assert!(
-        checkpoint_file.exists(),
-        "checkpoint file should be created at {}",
-        checkpoint_file.display()
-    );
-}
-
-/// Test 2: Session pool handles 429 responses gracefully.
-///
-/// The mock server returns 429 for the first request, then 200 for subsequent
-/// requests. The session pool should mark the domain as unhealthy after
-/// max_failures consecutive failures and skip it.
-#[ignore = "resurrected: pending triage"]
-#[tokio::test]
-async fn test_engine_with_session_pool_429() {
-    let server = MockServer::start().await;
-
-    // First request returns 429
-    Mock::given(method("GET"))
-        .and(path("/index.html"))
-        .respond_with(ResponseTemplate::new(429).set_body_string("Too Many Requests"))
-        .up_to_n_times(1)
-        .expect(1)
-        .mount(&server)
-        .await;
-
-    // Subsequent requests return 200 (but engine may not retry the seed)
-    Mock::given(method("GET"))
-        .and(path("/index.html"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_string("<html><body><h1>Hello</h1></body></html>"),
-        )
-        .mount(&server)
-        .await;
-
-    let config = test_config(&server.uri());
-    let options = EngineOptions {
-        checkpoint_path: None,
-        session_pool_enabled: true,
-        ignore_robots: true,
-        js_strategy: JsStrategy::Static,
-        autoscale_enabled: false,
-        ..Default::default()
-    };
-
-    // The crawl should NOT panic — 429 is handled gracefully.
-    // The engine may return an error (network error for 429) or succeed
-    // depending on retry logic, but it must not crash.
-    let result = crawl_site_with_options(config, options).await;
-
-    // We accept either Ok or Err — the point is no panic.
-    match result {
-        Ok(crawl_result) => {
-            // Seed returned 429 → engine may report 0 pages (graceful handling)
-            // or 1 page if it retried successfully. Both are valid.
-            println!(
-                "429 test: Ok with {} pages crawled",
-                crawl_result.total_pages
-            );
-        },
-        Err(e) => {
-            // 429 is a valid error — just verify it's a network-type error
-            let msg = e.to_string();
-            assert!(
-                msg.contains("429") || msg.contains("network") || msg.contains("rate limit"),
-                "error should mention 429 or network: {msg}"
-            );
-        },
-    }
-}
-
-/// Test 3: Engine resumes from an existing checkpoint.
-///
-/// Creates a checkpoint with one visited URL, then verifies the engine
-/// skips that URL and starts from the remaining queue.
-#[ignore = "resurrected: pending triage"]
-#[tokio::test]
-async fn test_engine_resume_from_checkpoint() {
-    let server = MockServer::start().await;
-
-    // Seed page with a link to /page2.html
-    Mock::given(method("GET"))
-        .and(path("/index.html"))
-        .respond_with(ResponseTemplate::new(200).set_body_string(
-            r#"<html><body>
-                    <a href="/page2.html">Page 2</a>
-                </body></html>"#,
-        ))
-        .mount(&server)
-        .await;
-
-    // page2 returns content
-    Mock::given(method("GET"))
-        .and(path("/page2.html"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_string("<html><body><h1>Page 2</h1></body></html>"),
-        )
-        .mount(&server)
-        .await;
-
-    let tmp = TempDir::new().unwrap();
-    let checkpoint_dir = tmp.path().join("checkpoints");
-    std::fs::create_dir_all(&checkpoint_dir).unwrap();
-
-    // Pre-create a checkpoint that marks the seed as already visited
-    use webfang::BincodeCheckpoint;
-    let seed_url = format!("{}/index.html", server.uri());
-    let mut visited = std::collections::HashSet::new();
-    visited.insert(seed_url);
-    let checkpoint = BincodeCheckpoint::from_state(&visited, &[], 1, vec![]);
-
-    let checkpoint_file = checkpoint_dir.join("crawl_checkpoint.json");
-    checkpoint.save(&checkpoint_file).unwrap();
-
-    // Now crawl with the same checkpoint dir — engine should resume
-    let config = test_config(&server.uri());
-    let options = EngineOptions {
-        checkpoint_path: Some(checkpoint_dir),
-        session_pool_enabled: false,
-        ignore_robots: true,
-        js_strategy: JsStrategy::Static,
-        autoscale_enabled: false,
-        ..Default::default()
-    };
-
-    let result = crawl_site_with_options(config, options).await;
-    assert!(
-        result.is_ok(),
-        "resume crawl should succeed: {:?}",
-        result.err()
-    );
-
-    let crawl_result = result.unwrap();
-    // The seed was already visited, so the engine should discover page2
-    // via the queue (if checkpoint restored it) or just finish quickly.
-    // The important thing is it doesn't re-crawl the seed.
-    println!(
-        "Resume test: crawled {} pages, {} total",
-        crawl_result.total_pages, crawl_result.total_pages
-    );
-}
-
-/// Test 4: robots.txt enforcement.
-///
-/// The mock server serves a robots.txt that disallows /private/.
-/// The engine with ignore_robots=false should respect it.
-#[ignore = "resurrected: pending triage"]
-#[tokio::test]
-async fn test_robots_txt_enforcement() {
-    let server = MockServer::start().await;
-
-    // robots.txt disallows /private/
-    Mock::given(method("GET"))
-        .and(path("/robots.txt"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_string("User-agent: *\nDisallow: /private/\n"),
-        )
-        .mount(&server)
-        .await;
-
-    // Seed page with links to allowed and disallowed paths
-    Mock::given(method("GET"))
-        .and(path("/index.html"))
-        .respond_with(ResponseTemplate::new(200).set_body_string(
-            r#"<html><body>
-                    <a href="/allowed.html">Allowed</a>
-                    <a href="/private/secret.html">Private</a>
-                </body></html>"#,
-        ))
-        .mount(&server)
-        .await;
-
-    // /allowed.html returns content
-    Mock::given(method("GET"))
-        .and(path("/allowed.html"))
-        .respond_with(
-            ResponseTemplate::new(200)
-                .set_body_string("<html><body><h1>Allowed</h1></body></html>"),
-        )
-        .mount(&server)
-        .await;
-
-    // /private/secret.html — should NOT be fetched if robots.txt is respected
-    Mock::given(method("GET"))
-        .and(path("/private/secret.html"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_string("<html><body><h1>Secret</h1></body></html>"),
-        )
-        .expect(0) // Must NOT be called
-        .mount(&server)
-        .await;
-
-    let seed = Url::parse(&format!("{}/index.html", server.uri())).unwrap();
-    let config = CrawlerConfig::builder(seed)
-        .max_depth(1)
-        .max_pages(10)
-        .delay_ms(1)
-        .concurrency(1)
-        .timeout_secs(5)
-        .ignore_robots(false) // Respect robots.txt
-        .build();
-
-    let options = EngineOptions {
-        checkpoint_path: None,
-        session_pool_enabled: false,
-        ignore_robots: false,
-        js_strategy: JsStrategy::Static,
-        autoscale_enabled: false,
-        ..Default::default()
-    };
-
-    let result = crawl_site_with_options(config, options).await;
-    assert!(
-        result.is_ok(),
-        "crawl with robots.txt should succeed: {:?}",
-        result.err()
-    );
-
-    let crawl_result = result.unwrap();
-    // Should have crawled index + allowed, but NOT /private/secret.html
-    assert!(
-        crawl_result.total_pages >= 1,
-        "should crawl at least the seed page"
-    );
-    // Verify the private page was not fetched (wiremock .expect(0) handles this)
-}
